git-submitd: allow users to push to the same repo, and keep track of what they pushed with tags

+++ cmd/git-submitd/submission-trigger.py
@@ -9,7 +9,7 @@ import time
 import shutil
 import argparse
 import tempfile
-from subprocess import check_call
+from subprocess import check_call, check_output, STDOUT
 
 parser = argparse.ArgumentParser()
 parser.add_argument('--repo')
@@ -17,28 +17,33 @@ parser.add_argument('--user')
 parser.add_argument('--remote')
 parser.add_argument('--key')
 parser.add_argument('--name')
+parser.add_argument('--tags')
 args = parser.parse_args()
+
+tags = args.tags.split("\x00")
 
 print
 print "Thanks for pushing some code!"
 print "==============================================================="
 print "You are user: %s" % args.user
-print "You pushed repo: %s" % args.repo
+print "You pushed to repo: %s" % args.repo
 print "You came from: %s" % args.remote
 print "The repo name is: %s" % args.name
 print "Your public key is: %s..." % args.key[:40]
+print "Tags pushed: %s..." % ", ".join(tags)
 print
 
-print "You pushed:"
-try:
-  worktree = tempfile.mkdtemp()
-  # git ls-tree -r is probably better than doing a checkout and then a find,
-  # but for demonstrative purposes, showing users how to get the working tree
-  # on disk seems useful.
-  check_call(["git", "--git-dir", args.repo, "--work-tree", worktree,
-              "checkout", "-f"])
-  check_call(["find", worktree])
-finally:
-  shutil.rmtree(worktree)
+if tags:
+  print "You pushed:"
+  try:
+    worktree = tempfile.mkdtemp()
+    # git ls-tree -r is probably better than doing a checkout and then a find,
+    # but for demonstrative purposes, showing users how to get the working tree
+    # on disk seems useful.
+    check_output(["git", "--git-dir", args.repo, "--work-tree", worktree,
+                "checkout", "-f", tags[0]], stderr=STDOUT)
+    check_call(["find", worktree, "-printf", "./%P\n"])
+  finally:
+    shutil.rmtree(worktree)
 
-print
+  print
